commit added API for fetching new Data
