+++ app/css/index.scss
@@ -3,3 +3,4 @@
 @import "./user-info";
 @import "./user-panel";
 @import "./modal-container";
+@import "./sort-feature.scss";
