+++ app/css/user-card.scss
@@ -1,7 +1,6 @@
 .user-card {
   display: flex;
   flex-flow: row nowrap;
-
   padding: 16px;
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.07);
   border-radius: 4px;
@@ -19,7 +18,7 @@
   }
 
   > .info {
-    flex: 2 0 66%;
+    flex: 2 0 46%;
     display: flex;
     justify-content: center;
     align-content: center;
