+++ app/css/user-gallery.scss
@@ -7,7 +7,11 @@
     grid-template-rows: auto;
     grid-gap: 16px;
 
-    @media screen and (min-width: 768.1px) {
+    @media screen and (min-width: 768px) {
+      grid-template-columns: 1fr 1fr ;
+    }
+
+    @media screen and (min-width: 900px) {
       grid-template-columns: 1fr 1fr 1fr;
     }
   }
