+++ app/css/user-info.scss
@@ -34,7 +34,7 @@
   }
 
   > .company > .name {
-    font-size: 18px;
+    font-size: 15px;
     font-weight: 600;
   }
