+++ package.json
@@ -13,6 +13,7 @@
     "@types/node": "20.4.2",
     "@types/react": "18.2.15",
     "@types/react-dom": "18.2.7",
+    "axios": "^1.7.4",
     "boring-avatars": "^1.10.1",
     "classnames": "^2.3.2",
     "eslint": "8.45.0",
@@ -23,6 +24,7 @@
     "react-dom": "18.2.0",
     "react-focus-lock": "^2.9.5",
     "react-icons": "^4.10.1",
+    "react-select": "^5.8.0",
     "typescript": "5.1.6"
   },
   "devDependencies": {
